test(topology): extract reusable topology change simulator
